1. Removed code for inclusion in a follow up PR.
