+++ iris-mpc-utils/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod constants;
-pub mod convertor;
 pub mod genesis;
 pub mod graphs;
 pub mod irises;
