+++ iris-mpc-utils/src/state/aws/mod.rs
@@ -1,2 +1,1 @@
 pub mod ops;
-mod utils;
