--- iris-mpc-utils/src/state/fsys/mod.rs
@@ -1,3 +1,0 @@
-pub mod local;
-pub mod reader;
-pub mod writer;
