+++ iris-mpc-utils/src/state/mod.rs
@@ -1,3 +1,2 @@
 pub mod aws;
-pub mod fsys;
 pub mod pgres;
